A TON of work on redoing the battle GUI completely with a thread.
Note: from here on out, every update will NOT be compatible with a previous save. This is due to a ton of engine changes and Data Structure changes that change the way some of the classes work with each other, and in the future, will also modify the Player's box fields so its a 2D array instead of 3 fixed boxes. Will post when saves will be safe to start in the future and won't be voided out.